ha: move dependencies to plugin
